feature(profile-photo-rekognition): refactor arch
